+++ mercearia-master/mercearia/src/br/com/mercearia/modelo/Fornecedor.java
@@ -4,45 +4,55 @@ public class Fornecedor {
 	private int id;
 	private String nome;
 	private String endereco;
-	private int telefone;
+	private long telefone;
 	private String email;
-	private Long cnpj;
-	
-	public Long getCnpj() {
+	private long cnpj;
+
+	public long getCnpj() {
 		return cnpj;
 	}
-	public void setCnpj(Long cnpj) {
+
+	public void setCnpj(long cnpj) {
 		this.cnpj = cnpj;
 	}
+
 	public int getId() {
 		return id;
 	}
+
 	public void setId(int id) {
 		this.id = id;
 	}
+
 	public String getNome() {
 		return nome;
 	}
+
 	public void setNome(String nome) {
 		this.nome = nome;
 	}
+
 	public String getEndereco() {
 		return endereco;
 	}
+
 	public void setEndereco(String endereco) {
 		this.endereco = endereco;
 	}
-	public int getTelefone() {
+
+	public long getTelefone() {
 		return telefone;
 	}
-	public void setTelefone(int telefone) {
+
+	public void setTelefone(long telefone) {
 		this.telefone = telefone;
 	}
+
 	public String getEmail() {
 		return email;
 	}
+
 	public void setEmail(String email) {
 		this.email = email;
 	}
-	
 }
